+++ components/shared/Tabs.module.scss
@@ -1,0 +1,112 @@
+@use "style/mixin";
+@use "style/function";
+@use "style/settings";
+
+$tabMaxNum: 10;
+
+.tab {
+  display: flex;
+  flex-wrap: wrap;
+  list-style: none;
+  border-top: 3px solid settings.$color-black;
+  border-right: 3px solid settings.$color-black;
+  @include mixin.lap {
+    display: inline-block;
+    border: none;
+    min-width: 675px;
+  }
+
+  // タブ最大10個まで重なり順の指定
+  @for $i from 1 through $tabMaxNum {
+    &__item {
+      &:nth-child(#{$i}) {
+        z-index: #{($tabMaxNum + 1) - $i};
+      }
+    }
+  }
+
+  &__item {
+    width: 50%;
+    border-left: 3px solid settings.$color-black;
+    border-bottom: 3px solid settings.$color-black;
+    @include mixin.lap {
+      display: inline-block;
+      width: inherit;
+      border: none;
+      text-align: center;
+      position: relative;
+    }
+
+    // タブが奇数個の場合最後の要素を全幅にする
+    &:nth-child(2n + 1) {
+      &:last-child {
+        width: 100%;
+        @include mixin.lap {
+          width: 154px;
+        }
+      }
+    }
+
+    a {
+      display: block;
+      padding: 1em;
+      position: relative;
+      @include mixin.lap {
+        // tabs
+        width: 154px;
+        line-height: 57px;
+        padding: 0;
+        background: url(data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTcxIiBoZWlnaHQ9IjY0IiB2aWV3Qm94PSIwIDAgMTcxIDY0IiBmaWxsPSJub25lIiB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciPgo8cGF0aCBkPSJNMS41IDEuNUgxNTguNzI2TDE2OC43MzQgNjIuNUgxLjVWMS41WiIgZmlsbD0iI2ZhZmFmYSIgc3Ryb2tlPSIjMEQwRDBEIiBzdHJva2Utd2lkdGg9IjMiLz4KPHBhdGggZD0iTTMgNTNMMTY1LjUgNTRMMTY3LjUgNjRIM1Y1M1oiIGZpbGw9IiNmYWZhZmEiLz4KPC9zdmc+Cg==)
+          no-repeat;
+        background-size: contain;
+        cursor: pointer;
+      }
+
+      &::after {
+        // arrow
+        content: "";
+        display: inline-block;
+        width: 1em;
+        height: 1em;
+        background: url(data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iOSIgaGVpZ2h0PSIxNCIgdmlld0JveD0iMCAwIDkgMTQiIGZpbGw9Im5vbmUiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyI+CjxwYXRoIGQ9Ik0xIDFMOCA2Ljk5N0wxIDEzIiBzdHJva2U9IiMwMjAxMEMiIHN0cm9rZS1taXRlcmxpbWl0PSIxMCIgc3Ryb2tlLWxpbmVjYXA9InJvdW5kIiBzdHJva2UtbGluZWpvaW49InJvdW5kIi8+Cjwvc3ZnPgo=)
+          center no-repeat;
+        background-size: contain;
+        position: absolute;
+        top: 50%;
+        right: 1em;
+        transform: translateY(-50%);
+        @include mixin.lap {
+          content: inherit;
+        }
+      }
+    }
+
+    &.isActive {
+      @include mixin.lap {
+        z-index: #{($tabMaxNum + 1)};
+      }
+      a {
+        font-weight: bold;
+        background-color: settings.$color-yellow;
+        @include mixin.lap {
+          width: 162px;
+          line-height: 60px;
+          background: url(data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMTc2IiBoZWlnaHQ9IjY2IiB2aWV3Qm94PSIwIDAgMTc2IDY2IiBmaWxsPSJub25lIiB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciPgo8cGF0aCBkPSJNMS41IDEuNUgxNjMuODg4TDE3NC4yMzQgNjQuNUgxLjVWMS41WiIgZmlsbD0iI0ZBRkFGQSIgc3Ryb2tlPSIjMEQwRDBEIiBzdHJva2Utd2lkdGg9IjMiLz4KPHBhdGggZD0iTTMgNTRIMTcxTDE3MyA2NkgzVjU0WiIgZmlsbD0iI0ZBRkFGQSIvPgo8cGF0aCBkPSJNMyAzSDE2Mi42MTNMMTY0LjQyIDE0SDNWM1oiIGZpbGw9IiNGRkY2NTciLz4KPC9zdmc+Cg==)
+            no-repeat;
+          background-color: inherit;
+          background-size: contain;
+        }
+
+        &::after {
+          transform: translateY(-50%) rotate(90deg);
+        }
+      }
+    }
+  }
+
+  &__item + &__item {
+    @include mixin.lap {
+      margin-left: -8px;
+    }
+  }
+}
